+++ packages/powersync-sdk-web/src/worker/WASQLiteDB.worker.ts
@@ -1,0 +1,4 @@
+import * as Comlink from 'comlink';
+import { _openDB } from './open-db';
+
+Comlink.expose(async (dbFileName: string) => Comlink.proxy(await _openDB(dbFileName)));
